Moved FakeResponseList into models for testing
