+++ src/test/java/lithium/university/TwitterCacheTest.java
@@ -1,5 +1,6 @@
 package lithium.university;
 
+import lithium.university.models.FakeResponseList;
 import org.junit.Before;
 import org.junit.Test;
 import twitter4j.ResponseList;
